+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - `Elements.Geometry.Profiles.WProfileType`
 - `Elements.Geometry.Profiles.WProfile`
 - `Elements.Geometry.Profiles.WProfileFactory`
+- `Grid2d.GetTrimmedCellProfiles`
+- `Ceiling`
 
 ### Changed